requirements.txt
asgiref==3.9.1
certifi==2025.8.3
channels==4.3.0
channels_redis==4.3.0
charset-normalizer==3.4.3
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
idna==3.10
msgpack==1.1.1
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytesseract==0.3.13
python-decouple==3.8
redis==6.4.0
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
